profiles: mask =dev-lang/python-3.12.8 and =dev-lang/python-3.13.1
3.12.8 and 3.13.1 break Mozilla's build system `mach` which is used in
firefox, thunderbird, and spidermonkey. It's not clear if it's a CPython
bug or a Mozilla bug (I've informed both sides) but the CVEs fixed in
these security releases are rather minor indeed, especially when weighed
against those packages being unbuildable.

LLVM 19 and Rust 1.82.0 got stabled last night together with a default
change to LLVM 19 which means more users will hit this too as they have
to rebuild FF and friends.

Done as lesser of two evils after discussion w/ mgorny.

Bug: https://bugs.gentoo.org/945857
Bug: https://bugs.gentoo.org/945882
Bug: https://bugs.gentoo.org/946002
Bug: https://bugs.gentoo.org/945850
Bug: https://bugs.gentoo.org/945849
Bug: https://bugs.gentoo.org/945845
Signed-off-by: Sam James <sam@gentoo.org>