dev-cpp/cpp-httplib: remove myself (xgqt) from maintainers
To ensure cpp-httplib does not break rdeps a lot to testing has to be
done with LLVM pkgs, as pointed out by mgorny. Currently my only use of
httplib was for openrgb and I was not checking out the LLVM
compatibility, it is related to also a dispute about whether httplib
SLOT should follow SONAME or the exact version designated by upstream.

Signed-off-by: Maciej Barć <xgqt@gentoo.org>