dev-java/openjdk: ensure libc++ is linked dynamically
OpenJDK attempts to link the C++ STL statically by default, which causes
linking to fail with libc++ due to some relocation errors. Pass
--with-stdc++lib=dynamic when libc++ is used to ensure that libc++ is
linked dynamically. I'm sure this issue also exists in OpenJDK 11 and
OpenJDK 8 but those versions are already mega broken with libc++, so
this would not be very helpful.

Closes: https://bugs.gentoo.org/935803
Signed-off-by: Violet Purcell <vimproved@inventati.org>