app-shells/starship: Use optfeature for fonts
These fonts seem to work decently well with starship in my experience,
but they are not the only suitable fonts. The user can install any
nerdfont they wish as described by uptsream's documentation.

starship doesn't require fonts to be installed on the same machine it is
running on necessarily. For example, if starship is installed on a
remote system accessed via ssh, the fonts should be installed on the
client machine.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>
Closes: https://github.com/gentoo/gentoo/pull/39636
Signed-off-by: Joonas Niilola <juippis@gentoo.org>