qt6-build.eclass: replace assert with pipestatus
Not very interesting to show pipestatus -v when the 2nd command
is just tail(1), realistically it's CC that will have failed.

Preemptively limit to EAPI=8 for when port to 9, albeit that will
likely be short lived as there is no need for extended EAPI=8 support
here after Qt ebuilds are migrated & stabilized.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>