llvm-core/clang: drop llvm_prepend_path use
The hackery with PATH here is wrong and it breaks building Clang itself
with a different version of Clang (which is why we're looking at rejigging
how the eclass works/replacing it with llvm-r1.eclass [0]).

As suggested by Matt, use LLVM_CMAKE_DIR instead so that Clang finds
the right version of LLVM to build against without any sort of nonsense
messing with build tooling.

[0] https://public-inbox.gentoo.org/gentoo-dev/d5489fa24ef3d1129540879e628120addb3af8ce.camel@gentoo.org/

Closes: https://bugs.gentoo.org/942314
Closes: https://bugs.gentoo.org/944788
Thanks-to: Matt Whitlock <gentoo@mattwhitlock.name>
Signed-off-by: Sam James <sam@gentoo.org>