dev-python/pyqt6-webengine: add 6.8.0
Almost identical to previous snapshot except for the removal of the
QVariant include which appears to be a mistake, quick restore it with
a patch for now.

No sip-6.9.1 PEP 625 names yet due to this being a "old" release
that was just pending being published on pypi, cleaning up will be
for the next one.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>