dev-python/pyqtgraph: bdepend on sys-apps/gawk
It turns out that `nawk` does not like the awk expressions
used in this ebuild.

Since it's not as portable as I'd hoped, longer term we
may want to replace the substitutions with a python
script or rework the `awk` script to be more POSIX-y.

Tidied up some `liborder` logic that still supported
the (never implemented) pyside6 USE.

Also stop explicitly excluding pyside2 tests as we rewrite
the test logic to only run tests for USE-enabled frontends.

Signed-off-by: Matt Jolly <kangie@gentoo.org>