x11-wm/fvwm3: stop forcing `-fsigned-char` on some arches
After discussion with upstream this is not required (and
is not used in e.g. Debian). It is likely that this was
to support a legacy fvwm and was carried forwards when
an early form of this ebuild was attached to the
bug requesting fvwm3.

Drop `flag-o-matic` as this is not required.

As the package is not yet keyworded on these arches,
and was typo'd to `-fsigned-chars` resulting in build
failures, we can safely assume that there are no
real-world customers - it is also very unlikely to hurt
anything.

Signed-off-by: Matt Jolly <kangie@gentoo.org>