x11-drivers/nvidia-drivers: update DRM_*_HELPERS error text
I915 now selects TTM in newer kernels making it usable, however
the text is unaware of changes and versions, so just drop the
mention that I915 is not usable and suggest checking the kernel's
help.

Also list QXL example first, it's smaller to build. Still keep
AMDGPU listed mostly because it's more recognizable by users.

Leaving 535 as-is, last I know of TTM_HELPERS doesn't matter for
it and it likely does not have that long of support.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>