dev-util/rizin: depend on libpcre2[jit], force bundled tree-sitter-c
Rizin can automagically pick up a dependency on an incompatible
system-tree-sitter-c, so instead force usage of the forked
subproject. Note that I'm patching this as a one-off in this version
since the forked package has been renamed upstream and thus this
should be obsolete by the next release.

Also add a dependency on libpcre2[jit], since Rizin automagically
enables usage of it based on the target architecture, and we're not
keyworded for anywhere that doesn't have support except for x32, for
which usage is likely exceedingly rare (if existent).

Bug: https://bugs.gentoo.org/943666
Closes: https://bugs.gentoo.org/945113
Signed-off-by: John Helmert III <ajak@gentoo.org>