llvm-runtimes/libunwind: fix build with GCC
libunwind's build system tests whether ${CXX} supports -nostdlib++ but
then passes that flag to ${CC}, which may not support it (and does not
if it is GCC). Work around this by deleting upstream's hack that forces
libunwind to be linked using ${CC} rather than ${CXX} despite that it
contains C++ objects.

Signed-off-by: Matt Whitlock <gentoo@mattwhitlock.name>