llvm-runtimes/libcxx: set CMAKE_PREFIX_PATH instead of munging PATH
Avoid calling llvm_prepend_path or llvm_pkg_setup, as those functions
permute PATH and can change which version of LLVM+Clang is referenced by
CC, CXX, et al. It is okay (and necessary in some cases) to build older
versions of LLVM dependents using the latest installed version of Clang
as long as CMake is told where to find the applicable LLVM installation.

Additionally, relax BDEPEND on llvm-core/clang when USE="clang" since
any newer version of Clang can be used to build; it doesn't have to be
the exact same version as the version of the package being built. All
that matters is that the correct headers and libraries are used, and
that's accomplished by setting CMAKE_PREFIX_PATH.

Bug: https://bugs.gentoo.org/942314
Signed-off-by: Matt Whitlock <gentoo@mattwhitlock.name>