rust.eclass: be verbose when checking if a Rust is suitable
This commit enables the Rust eclass to provide feedback to users
on what it's actually doing and why a particular Rust was
(or was not) deemed suitable for use.

To do this we now:

- Explicitly note if a Rust slot was skipped due to LLVM_SLOT
  incompatibility
- Provide a python-utils-r1 style `Checking whether Rust SLOT is
  suitable ...` for each slot
- List each package (and usedep) that we are checking for
- Provide an error message that clearly explains the requirements
  if no suitable package is available.

`_get_rust_slot` has been adjusted to export RUST_SLOT and RUST_TYPE
if a suitable slot is found.

Still TODO is enhancing output for `rust_check_deps` (perhaps
ebegin/eend?), however there don't appear to be any consumers
in the wild, this can probably wait a little longer.

Signed-off-by: Matt Jolly <kangie@gentoo.org>