dev-util/cargo-c: require static-libs for sys-libs/zlib
I am not sure what causes this requirement. It doesn't appear to be from
libz-sys as nothing should be setting LIBZ_SYS_STATIC or enabled the
static feature for the crate.

As a bonus also run cargo_src_configure. It wont't affect anything as
not features are enabled/disabled, but this means a user could use
ECARGO_EXTRA_ARGS if they find a need for it.

Bug: https://bugs.gentoo.org/938018
Signed-off-by: Alfred Wingate <parona@protonmail.com>