x11-drivers/nvidia-drivers: update 390/470 support year
Was announced that LTS kernel 6.1 will get an extra year of
support, so this may give an extra year in the tree to these
drivers (still depends on if Gentoo's kernel maintainers are
going to keep these in the tree for that long, or if these
drivers will still be in a usable state by then).

Given 6.6.x hasn't been extended, 470 users may need to go
back to 6.1.x for that extra year (without patching anyway).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>