dev-qt/qtcore: unkeyword ~sparc
Qt6 has a lot of odd issues on sparc which likely exist on Qt5
as well hidden behind tests being disabled.

...and then USE=qt5 been masked since 2015 and nobody complained.
Without that USE (e.g. USE=gui) there is only 5 revdeps that it is
doubtful many care for on sparc (e.g. cmake[gui]). Indicates that
there is little interest and probably no real users on Gentoo.

So, to simplify migration to Qt6 (which is also dropping sparc),
dropping all sparc keywords at once for dev-qt/* and revdeps.

This could be revisited if we find actual Qt sparc users.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>