x11-drivers/nvidia-drivers: add comment about kernel 6.13 & fbdev=1
NVIDIA has fixed 550.142 to build with 6.13-rc3 (565.77 fails), but
has not addressed/noticed that fbdev=1 is broken. So keep a reminder
to not add =6.13 even if it builds when 6.13 releases.

Also note in 565.77 in case a new version releases that only fixed
the build failure and not fbdev=1.

Patches exists, but would rather not use these, hard to tell if
they're correct. Users should stick to 6.12 or older until proper
fixes.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>