sys-devel/gcc: keyword 14.2.1_p20241221
Has a bunch of misc. fixes but importantly fixes a miscompilation with
3DNow! instructions where x87 FPU state was left corrupted. This only
affected >= GCC 14 and is now fixed.

Unfortunately, the nature of the bug means that all packages may need
to be recompiled (see https://bugs.gentoo.org/942573#c76 for more detail
there). We can still consider a news item describing how to find potentially
affected packages, but that's not a reason to put off keywording (and shortly,
stabilisation).

Thanks again to amonakov for the help in debugging, jospezial for the report,
immolo for initially working with amonakov on it, and all others who
helped & offered help. And Uros for fixing it upstream, of course.

Will file a stable bug soon. I'd been planning on keywording this
shortly anyway but was waiting for Christmas for things to settle down:
now is a good time, and also was prompted by a potential other report
of this on the forums at https://forums.gentoo.org/viewtopic-p-8849550.html.

Bug: https://gcc.gnu.org/PR117926
Bug: https://bugs.gentoo.org/942573
Signed-off-by: Sam James <sam@gentoo.org>