games-util/xpadneo: sync live
Spotted that 0.9.7 renamed a rules file that we weren't installing.
May as well just use a wildcard to avoid missing these changes (not
a super important file though, it's a workaround for streamlink).

Alternatively could consider building+installing using upstream's
Makefile which been improved in live, albeit it wouldn't simplify
the ebuild much and it installs udev files to /etc rather than
/usr/lib.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>