dev-qt/qtwebengine: adjust system ffmpeg comment
Seeing distros that use system's ffmpeg (like Arch) hit obscure
issues while it works on Gentoo. Like some mp3 or opus files not
playing. In case of mp3, Qt had to do a workaround for system
ffmpeg to ensure it picks the right mp3 decoder given chromium
does not support the others. Opus case is unclear still.

If we ever switch to system (patched) ffmpeg (which would be nice
on paper esp. for binpkg due to USE=bindist), it should probably
use extra consideration and also be kept optional to debug.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>