dev-perl/Cache-Memcached-Fast: fix build failure with LTO
This removes now-unneeded patch that removes '-flto' from CFLAGS, but leaves the
job specifier as-is. As of now, Cache-Memcached-Fast builds just fine with LTO
enabled.

For instance, if '-flto=16' is present in CFLAGS, the patch would only remove
the '-flto', leaving '=16' as an argument to a compiler. '=16' not being a valid
option, results in a build failure.

Closes: https://bugs.gentoo.org/946880
Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/39837
Closes: https://github.com/gentoo/gentoo/pull/39837
Signed-off-by: Sam James <sam@gentoo.org>