games-strategy/ja2-stracciatella: remove unused fltk dep
Was going around fltk revdeps adding binding := (soname changed)
only to find that this does not use fltk with -DBUILD_LAUNCHER=OFF.

Doesn't link with it when it's installed, and builds fine when it's
not.

fwiw left the LOCAL_FLTK_LIB=OFF alone to ensure it doesn't fallback
to bundled, not that this code path is reached.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>