net-misc/tigervnc: add upper <fltk-1.4 bound for now
Fails to build. Meant to fix it, but seen a comment by upstream
saying that the build failure may as well be a good thing given
it's incompatible with fltk-1.4 implying it likely has other issues.

Also add binding operator given gained a subslot & soname changed,
not that it matters with the upper bound.

Skipping upper bound from 9999, given upstream is already aware
I'd imagine it'll be fixed in the near future.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>