media-sound/zynaddsubfx: add binding := to fltk & fix build with 1.4
1.4.x seems to have dropped some transitive includes, and this needs
to include what it uses.

:= due to new soname

Have not tried wayland (new in fltk-1.4), but given the build fix
revealed that it uses some X-specific bits, odds are it'll be
broken without fltk[X(+)] and so added that as well.

~arch version dropped fltk so this is just for stable, but had to
do a no-op ~arch revbump only to prevent the stable revbump from
becoming a higher revision.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>