net-wireless/limesuite: add binding := to fltk, soname changed
For what it's worth anyway, ebuild appears broken for other reasons at
the moment (not planning to look into it myself, one *could* just
remove the offending `rm` line but issue *may* be that the GUI is
getting disabled for some reason and then it skips Desktop files).
Skipping the revbump given a rebuild would likely fail anyway.

Bug: https://bugs.gentoo.org/922785
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>