dev-scheme/guile-lib: add 0.2.8.1
This includes Zack Weinberg's patch to fix tests w/ guile-3.0, which
I've only applied for 3.0.

But in the course of spending a while to get the ebuild working with
two impls and only patching one, I found that 2-2 was clearly not cared
for upstream anymore as I hit two different issues and one of them
had an unreplied-to patch on the ML, so I gave up. Just declare compat
w/ 3-0 and worry about older later for this version on the offchance
someone requests it.

Closes: https://bugs.gentoo.org/877785
Closes: https://bugs.gentoo.org/941404
Signed-off-by: Sam James <sam@gentoo.org>