dist-kernel-utils.eclass: implement basic support for install with ROOT
Before this change the kernel installation is silently skipped in all cases.
Using sys-kernel/installkernel to generate an initramfs/uki will indeed be
a huge mess so when an initramfs is requested we now print a helpful error
message. If we do not want an initramfs, or the initramfs/uki was generated
with portage using USE=generic-uki then we can safely try to install the kernel.

Installkernel-50 is capable of dealing with this situation in most cases, but it
will print a bunch of warnings since plugins are skipped. Bootloader
configuration, for example, cannot be updated in this case.

For kernel modules we can also safely make this change, all modules that can be
in the initramfs have the initramfs flag and if it is enabled then we will also
hit this error message.

End result of this change is more verbosity when kernel installation is skipped
and basic support for installing the kernel with --root set for simple
configurations (no initramfs or generic-uki with compat layout or uki layout).

Signed-off-by: Nowa Ammerlaan <nowa@gentoo.org>
Closes: https://github.com/gentoo/gentoo/pull/38842
Signed-off-by: Michał Górny <mgorny@gentoo.org>