sys-libs/glibc: Simplify and fix kernel header location logic
ALT_HEADERS dates back to 2004 [1] and ALT_BUILD_HEADERS back to 2007
[2], both before crossdev was born. ALT_BUILD_HEADERS was added to
crossdev in 2016 but then removed again in 2017. None of this seems
necessary anymore.

It's unclear exactly which scenario the linux/version.h logic was for, but this
also appears obsolete now. It was referred to as a "fallback case" at the time
it was added.

Now that we have ESYSROOT, we can simply rely on that instead. Tested with a
prefixed "Canadian Cross".

[1] https://gitweb.gentoo.org/archive/repo/gentoo-2.git/commit/?id=5d0403b9db95bfe58e9196c31997eedc97869881
[2] https://gitweb.gentoo.org/archive/repo/gentoo-2.git/commit/?id=b127166cfbce8930aac4ed893df23c3e28b24737

Closes: https://github.com/gentoo/gentoo/pull/39914
Signed-off-by: James Le Cuirot <chewi@gentoo.org>