zig.eclass: use absolute prefix + DESTDIR
Based on ghostty's packaging guidelines this appears to be the correct
way of imaging Zig packages. This makes it so that "zig build" always
has an absolute path to install files to. The trailing slashes from
prefix paths have been removed to follow the path building convention
of ${PREFIX}/other/path.

Signed-off-by: sin-ack <sin-ack@protonmail.com>