zig-utils.eclass: always set a DESTDIR for ezig
DESTDIR is prepended to the install prefix in order to install a copy of
the package to another path. This needs to always be present because
Zig build steps may contain install steps, and tests can depend on such
steps. We designate a default install directory under ${T} called
ZIG_STAGED_DESTDIR which is used by default. The real DESTDIR is then
overridden in zig_src_install in order to install to ${D}.

Signed-off-by: sin-ack <sin-ack@protonmail.com>