dev-util/sysprof: fix nastily propagating automagic dep
sysprof 45.0 and on introduce a systemd feature (journald integration)
which is implemented by depending on libsystemd.so if it is installed at
the time of building.

https://github.com/GNOME/sysprof/commit/88d0589fea80307fe8d49a9c4f93ef99cb6a5e8b

This dependency was never added during the version bump in ::gentoo. It
breaks, among other things, binhost caching. Fix by introducing
USE=systemd support to this package. We use strong blockers to prevent
systemd from even being installed when src_compile'ing with -systemd,
since otherwise it genuinely miscompiles. This is probably fine, since
people using systemd profiles have USE=systemd by default unless they go
out of their way to unset it, and people with non-systemd profiles
shouldn't have this USE set at all.

(Note that it's fine to install a binpackage built without systemd
support, even on a systemd profile.)

Straight to stable, the package nastily breaks on openrc systems and the
changes are purely to dependency metadata.

Fixes: 29d274e3ecdde94273ddd9f247efa9c5e12dab73
Fixes: https://github.com/gentoo/gentoo/pull/34771
Closes: https://bugs.gentoo.org/947256
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>