zig-utils.eclass: add ZIG_NEEDS_LLVM variable
If set to a non-empty value, the package will BDEPEND on a Zig package
with LLVM enabled. This is currently required for packages that require
C/C++ source files to be compiled with Zig.

There's some special logic required to detect which Zig binary we should
actually use, because we may have dev-lang/zig[-llvm] installed with
dev-lang/zig-bin (in which case we would need to pick zig-bin over zig).

Tested in a clean stage3 chroot with dev-lang/zig[-llvm] and
dev-lang/zig-bin installed, in which case zig-utils.eclass will pick
zig-bin over zig.

Co-authored-by: Eric Joldasov <bratishkaerik@landless-city.net>
Signed-off-by: sin-ack <sin-ack@protonmail.com>