sys-libs/glibc: Fix cross-compiling glibc when clang is selected
binutils-config returns the wrong binary path for cross-compilers unless you
specify both the CHOST (actually CBUILD in this case) and CTARGET.

Cross-compilers do not have the bare binary names (e.g. gcc) in the binary path
returned by gcc-config, but the tuple-prefixed names are always present.

Tested with a cross build and a native build.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>