dev-games/godot: define mbedtls slot
No need for a revbump given the previous upper bound accomplished
the same job and this is a long rebuild, this is just about showing
slot awareness.

On a side-note, when we want to use mbedtls:3, will likely need to
update detect.py to use the new pkg-config files.

Per ebuild comment, also note that this currently bundles mbedtls3
but builds it differently than system's which is currently required
for compatibility. So using 3 is not an option at the moment either
way unless we switch to bundled, if support for 2 is dropped and that
is not resolved we may need to go ahead and use bundled.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>