mail-mta/proton-mail-bridge: fix src_test
The test change was wrong. Tests are restricted but if you unrestrict,
the default doesn't work, because it uses cmake_src_test which finds nothing. Need
to keep the explicit emake test call to use the test target from the Makefile instead.

Also, use -Onone because of what the targets in the Makefile do here
to avoid no output for long stretches of the build.

Signed-off-by: Sam James <sam@gentoo.org>