dev-build/libtool: silence maintainer mode noise
For 2.5.4, the only difference is two files being removed from a dist
check in libltdl/Makefile.in which is harmless. We already need autoconf
for this (we run eautoreconf, after all) and it also actually persists despite
experimenting with regenerating locally and also just letting it Do Its
Thing with maintainer mode (it'll continue trying to!), so just ignore it.

Closes: https://bugs.gentoo.org/940302
Signed-off-by: Sam James <sam@gentoo.org>