dev-qt/qtmultimedia: fix build w/ USE="gstreamer pulseaudio -vulkan"
QVersionNumber is only used with USE="gstreamer pulseaudio", but the
required header is not included for it. Seems works with USE="vulkan"
given Qt vulkan headers just happen to include it.

Furthermore only fails with clang, albeit haven't looked for what is
being treated differently with it (perhaps a version check somewhere
includes the header with gcc).

Closes: https://bugs.gentoo.org/947606
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>