kernel-build.eclass: resolve configure problem with USE=secureboot
if no keys set, then secureboot.eclass now inherits linux-info.eclass to check
for potential key candidates in the kernel config. This would set KV_FULL
and create a mess later on in the configure phase. Ensure that KV_FULL remains
local to the setup phase specifically for the kernel packages.

Signed-off-by: Nowa Ammerlaan <nowa@gentoo.org>