eclass/dotnet-pkg-base.eclass: add DOTNET_VERBOSITY user variable
DOTNET_VERBOSITY controls dotnet restore/build/test verbosity.
We allow the user to set this value to any other to value because it
might be helpful when debugging dotnet processes. On the other hand,
generally warnings and errors provide enough info for debugging package
emerge process.

Normally we use the default value - "minimal", that still reports what is
being built and any warnings/errors but it might come handy to set it
to a higher value for the Gentoo Tinderbox CI.

Signed-off-by: Maciej Barć <xgqt@gentoo.org>