dev-libs/libratbag: check if CONFIG_HIDRAW is set
libratbag uses `libudev`, namely devices in: `/dev/hidraw*`.  If
CONFIG_HIDRAW is not set, the package will never find any devices.
(Which was true in my case, as I had HIDRAW disabled.)

This is visible in the libratbag git commit df3c73e, function
`ratbagd_run_enumerate`, on line 436:

    r = udev_enumerate_add_match_subsystem(e, "hidraw");
    if (r < 0)
            goto exit;

This will not `goto exit` when there are no devices, neither does any of
the subsequent code. The code's intent is not wrong, if there really are
no HIDRAW-capable devices connected, it would work as if CONFIG_HIDRAW
is not set.

But it does purely depend on the fact that CONFIG_HIDRAW is set in the
first place, or else it just makes no sense to use libratbag and may
cause frustration if the intent is to really make use of libratbag (e.g.
piper) and won't ever find any devices.

Signed-off-by: Raphael Seitz <raphael.seitz18@gmail.com>