sys-libs/pam: stub out xmllint/xmlcatalog use
We don't have docbook5 packaged and it appears to have A Whole History
with a blogpost dating back to even 2008 from flameeyes. Anyway, it's
only needed for validation of output here, so we can stub it out and
stick our heads in the sand.

Thanks to ulm for the help.

Closes: https://bugs.gentoo.org/913087
Thanks-to: Ulrich Müller <ulm@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>