llvm-runtimes/libgcc: New package, 19.1.7 and 20.x
llvm-libgcc[0] is the LLVM subproject which provides a link script,
which can be used to build a library linked to compiler-rt and
llvm-libunwind, which serves as a full replacement of libgcc_s.

The main purpose of having it in Gentoo is ability to run binaries,
linked to libgcc_s, on musl-llvm systems.

[0] https://github.com/llvm/llvm-project/tree/main/llvm-libgcc

Bug: https://bugs.gentoo.org/946486
Signed-off-by: Michal Rostecki <vadorovsky@protonmail.com>
Pull-Request: https://github.com/gentoo/gentoo/pull/39868
Signed-off-by: Michał Górny <mgorny@gentoo.org>