cargo.eclass: Emit a warning if the package uses 300+ crates
Emit a QA warning suggesting the use of crate tarball, when the package
in question uses 300 crates or more.  Such a long crate lists cause
ebuilds and Manifests to grow very fast, causing significant space
consumption on end user systems (including users who are not using
the package in question) and git history growth.  On top of that,
fetching that many crates takes significant time.

The number of 300 is pretty arbitrary, chosen approximately to match
Manifests that are over 100 KiB in size.  We should probably look into
lowering in the future, as more packages are transitioned.

Closes: https://github.com/gentoo/gentoo/pull/40096
Signed-off-by: Michał Górny <mgorny@gentoo.org>