dev-python/vdf: Bump to 4.0
The original repository in the ValvePython GitHub group
is abandoned and hasn't seen any commit in the master
branch for three years.

Solstice Game Studios has more or less taken over the
maintenance of the project and is now providing a new
4.0 release. This release is crucial since the format
of the .vdf files that the current stable Steam client
is using has changed significantly. The old 3.4 release
can no longer parse these files, i.e. the release is
non-functional when using the stable Steam client.

Signed-off-by: Tobias Jakobi <tjakobi@math.uni-bielefeld.de>