www-client/chromium: depend on compiler-rt-sanitizers[cfi]
We enable CFI with USE=official when LTO is enabled. While most
profiles enable USE=cfi anyway, we can be more specific and ensure
that the USE is enabled when CFI is likely to be enabled - we can't
do that based on LTO so USE=official will serve as a reasonable
proxy.

Since llvm/runtimes/compiler-rt-sanitizers[cfi] is not available
on ppc64, special-case that arch and do not enable CFI.

Closes: https://bugs.gentoo.org/946853
Signed-off-by: Matt Jolly <kangie@gentoo.org>