net-libs/nodejs: Allow to use atomic builtins from compiler-rt instead of libatomic
Currently, the nodejs ebuild has a hard dependency on GCC, because of
dependency on atomics.

However, we provided a possibility to build compiler-rt with all atomic
builtins[0], which means that we don't have to require GCC when
compiler-rt can be used instead.

To make use of it, provide the `llvm-atomic-builtins` USE flag which
adds the `--use-compiler-rt-atomics` configure flag, telling the build
system to not link libatomic. That new USE flag is:

* Enabled by default in the ebuild.
* Masked in the base profile.
* Unmasked in the LLVM feature profile.

Which practially means, that it's disabled on all non-LLVM profiles and
enabled by default on LLVM profiles.

[0] https://github.com/gentoo/gentoo/commit/63b4ae7aaa6e520706e1237b649d8fe29f5aba83

Bug: https://bugs.gentoo.org/911340
Signed-off-by: Michal Rostecki <vadorovsky@protonmail.com>