net-misc/rsync: drop useless dependency on sys-apps/attr
When USE=xattr is enabled, the configure script checks for -lattr and
attempts to link to it for good luck. It also checks for attr/xattr.h,
but only includes it in an if/elif when sys/xattr.h doesn't exist --
this makes sense, since the former simply #includes the latter and then
emits a #warning telling you to migrate to libc. Presumably, this is
meant as a sort of infinite compatibility thing and "it's fine if you
use --as-needed" in practice.

rsync doesn't actually include any of libattr's non-deprecated headers,
just xattr.h and at runtime libattr.so is not needed despite
automagically detecting it during configure.

Upstream fix for false detection: https://github.com/RsyncProject/rsync/pull/753

Fixes:
```
 * VDB: detected possibly incorrect RDEPEND (net-misc/rsync-3.3.0-r2)
 * sys-apps/attr <
```

No revbump, this package won't be depcleaned anyway so forcing a rebuild
feels pointless.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>