net-misc/trurl: Install from release tarball instead of git tag
Since version 0.14, you are supposed to install trurl from the release
tarball instead of just using the source code at the git tag of the
release. The release tarball includes a precompiled trurl.1 man page
that is not included in git tree, and is not built by trurl because it
requires a script from the curl git repo.

I had to workaround the "trurl-0.16.tar.gz" distfile already existing on
mirrors for the git tag tarball
https://github.com/curl/trurl/archive/refs/tags/trurl-0.16.tar.gz
renaming the distfile "trurl-0.16-tmp.tar.gz"; this can be removed at
the next release.

Ref: https://github.com/curl/trurl/issues/380
Signed-off-by: Emanuele Torre <torreemanuele6@gmail.com>