app-emulation/dxvk: drop 1.10.3-r1
Kept this version for longer due to being the last version to support
legacy nvidia-drivers and older wine (plus being the last of the 1.x
in case 2.x introduced regressions), but if it fails to build then
that's about as far as it goes given won't patch this further.

If really need it, I'd suggest using winetricks to install prebuilts.

Closes: https://bugs.gentoo.org/948435
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>