app-arch/sharutils: update EAPI 7 -> 8, fix modern C issues
Patch removes which from m4 macros and regenerates configure.
Patch adds include of lib/gettext.h into only place it can go
so it works in glibc and musl - in the middle of *-opts.h file,
after condifional declaration of gettext or inclusion of libintl,
but before gettext is actually used.
In other places funky macro dance either undefines gettext or
uses it before marking it with asm attributes.
This supercedes other PR, and also lets all code to be aware of all
the defines from config.h and friends.

Closes: https://bugs.gentoo.org/941724
Closes: https://bugs.gentoo.org/894322
Bug: https://bugs.gentoo.org/900248
Closes: #31725
Signed-off-by: NHOrus <jy6x2b32pie9@yahoo.com>